+++ packages/pluggableWidgets/datagrid-web/src/components/StickyHeaderTable.tsx
@@ -1,5 +1,9 @@
 import { createElement, ReactElement, useCallback, useEffect, useRef, PropsWithChildren, Fragment } from "react";
 
+// this wrapper component works by adding "sticky-sentinel" div (and hides it using margin-bottom -1px)
+// in order to detect table going out of viewport
+// this allows position: sticky to only applies when container is "stuck"
+// this way we allow any absolute position to still be on top when sticky not "stuck"
 export function StickyHeaderTable(props: PropsWithChildren): ReactElement {
     const headerContainer = useRef<HTMLDivElement>(null);
     const observerCallback = useCallback(
@@ -29,7 +33,7 @@ export function StickyHeaderTable(props: PropsWithChildren): ReactElement {
     }, [headerContainer]);
     return (
         <Fragment>
-            <div className="sticky-sentinel" ref={headerContainer} style={{ height: 1 }}></div>,
+            <div className="sticky-sentinel" ref={headerContainer} style={{ height: 1, marginBottom: "-1px" }}></div>
             <div className="table" role="table">
                 {props.children}
             </div>