+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
--   We changed the DOM Structure for dropdown filter to appear inline with .react-datepicker component in order to make the component more accessible.
+-   We changed the DOM Structure for dropdown filter to appear inline with the container in order to make the component more accessible.
 
 ## [2.4.3] - 2023-06-21